refactor: 클래스 분리, 코드 포맷팅
