+++ src/main/kotlin/minesweeper/domain/Cell.kt
@@ -1,0 +1,10 @@
+package minesweeper.domain
+
+data class Cell(
+    val coordinate: Coordinate,
+    val hasMine: Boolean = false,
+) {
+    fun isItCoordinate(coordinate: Coordinate): Boolean {
+        return this.coordinate.itIt(coordinate)
+    }
+}
