+++ src/main/kotlin/minesweeper/domain/Coordinate.kt
@@ -1,0 +1,7 @@
+package minesweeper.domain
+
+data class Coordinate(val axisX: Int, val axisY: Int) {
+    fun itIt(coordinate: Coordinate): Boolean {
+        return this == coordinate
+    }
+}
